requirements.txt
langchain-core>=0.3.0
langchain-ollama>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=1.0.0
python-dotenv>=1.0.1
Pillow>=10.0.0
torch>=2.0.0
open-clip-torch>=2.24.0
pypdf>=4.0.0